requirements.txt
fastapi==0.123.0
uvicorn[standard]==0.38.0

SQLAlchemy==2.0.44
alembic==1.17.2
psycopg[binary]==3.3.0
asyncpg==0.29.0
pydantic==2.12.5
pydantic-settings==2.6.1

redis==5.0.7
passlib[bcrypt,argon2]==1.7.4
PyJWT==2.8.0
cryptography==43.0.1
python-multipart==0.0.6

requests==2.32.3
uvloop==0.19.0
pytest-asyncio==0.23.7
httpx==0.28.1
celery[redis]==5.3.6

aiosmtplib==5.0.0
